# schemas/openapi.yaml
openapi: 3.1.0
info:
  title: Evidence-Grounded Research Workbench API
  version: 0.2.0
  description: Full MVP API contract for the evidence-grounded multi-model long-horizon research workbench.
servers:
  - url: http://localhost:4000
    description: Local Fastify API
...
paths:
...
  /projects/{projectId}/evidence:
    get:
      summary: List evidence
      tags: [Evidence]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Evidence list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceListResponse'
    post:
      summary: Add manual evidence
      tags: [Evidence]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEvidenceRequest'
      responses:
        '201':
          description: Evidence created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceResponse'
  /evidence/{evidenceId}:
    get:
      summary: Get evidence by ID
      tags: [Evidence]
      parameters:
        - $ref: '#/components/parameters/EvidenceId'
      responses:
        '200':
          description: Evidence
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update evidence fields or status
      tags: [Evidence]
      parameters:
        - $ref: '#/components/parameters/EvidenceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEvidenceRequest'
      responses:
        '200':
          description: Evidence updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceResponse'
  /claims/{claimId}/search-evidence:
    post:
      summary: Search supporting evidence for claim
      tags: [Evidence]
      parameters:
        - $ref: '#/components/parameters/ClaimId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchEvidenceRequest'
      responses:
        '201':
          description: Evidence search completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceListResponse'
...
  /evidence/{evidenceId}/assess:
    post:
      summary: Run evidence assessment by selected models
      tags: [Evidence]
      parameters:
        - $ref: '#/components/parameters/EvidenceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssessEvidenceRequest'
      responses:
        '201':
          description: Evidence assessments created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceAssessmentListResponse'
  /projects/{projectId}/evidence/assess-pending:
    post:
      tags: [Evidence]
      summary: Assess all pending evidence for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string, format: uuid }
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pending evidence assessed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      assessed: { type: integer }
        '404':
          description: Project not found
  /projects/{projectId}/evidence/find-counter-evidence:
    post:
      tags: [Evidence]
      summary: Search for counter-evidence for project claims
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string, format: uuid }
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Counter-evidence search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      counterEvidenceFound: { type: integer }
        '404':
          description: Project not found
...
  /projects/{projectId}/evidence/quality:
    get:
      tags: [Evidence]
      summary: Get evidence quality assessment overview
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Evidence quality overview
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      totalCount: { type: integer }
                      assessedCount: { type: integer }
                      reliabilityDistribution:
                        type: object
                        additionalProperties: { type: integer }
                      sourceTypeDistribution:
                        type: object
                        additionalProperties: { type: integer }
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/evidence/stale:
    get:
      tags: [Evidence]
      summary: Get evidence items at risk of staleness
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Stale evidence list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/StaleEvidenceReport'
        '404':
          $ref: '#/components/responses/NotFound'
  /evidence/{evidenceId}/verify:
    post:
      tags: [Evidence]
      summary: Verify a piece of evidence's source integrity
      parameters:
        - name: evidenceId
          in: path
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      verified: { type: boolean }
                      message: { type: string }
        '404':
          $ref: '#/components/responses/NotFound'
  /evidence/{evidenceId}/provenance:
    get:
      tags: [Evidence]
      summary: Get evidence chain of provenance (how it was discovered, assessed, used)
      parameters:
        - name: evidenceId
          in: path
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Evidence provenance
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    additionalProperties: true
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Export Extras ───────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string }
...
    ClaimId:
      name: claimId
      in: path
      required: true
      schema: { type: string }
    EvidenceId:
      name: evidenceId
      in: path
      required: true
      schema: { type: string }
...
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
...
    Evidence:
      type: object
      required: [id, projectId, title, sourceType, reliability, relevance, status, stalenessRisk, createdAt]
      properties:
        id: { type: string }
        projectId: { type: string }
        claimId: { type: string, nullable: true }
        discoveredByModelId: { type: string, nullable: true }
        sourceUrl: { type: string, nullable: true }
        title: { type: string }
        publisher: { type: string, nullable: true }
        publishedAt: { type: string, format: date-time, nullable: true }
        retrievedAt: { type: string, format: date-time }
        sourceType:
          type: string
          enum: [official, academic, government, company, news, benchmark, blog, forum, user_input, unknown]
        excerpt: { type: string, nullable: true }
        summary: { type: string, nullable: true }
        rawContentRef: { type: string, nullable: true }
        reliability:
          type: string
          enum: [pending, high, medium, low, unusable]
        relevance:
          type: string
          enum: [pending, direct, indirect, weak, irrelevant]
        status:
          type: string
          enum: [pending_review, accepted, accepted_with_caution, rejected, irrelevant, needs_better_source]
        stalenessRisk:
          type: string
          enum: [low, medium, high]
        createdAt: { type: string, format: date-time }
    EvidenceAssessment:
      type: object
      required: [id, evidenceId, reviewerModelId, reliability, relevance, interpretationVerdict, notes, finalVerdict, createdAt]
      properties:
        id: { type: string }
        evidenceId: { type: string }
        reviewerModelId: { type: string }
        reliability:
          type: string
          enum: [high, medium, low, unusable]
        relevance:
          type: string
          enum: [direct, indirect, weak, irrelevant]
        interpretationVerdict:
          type: string
          enum: [correctly_used, overstated, misinterpreted, out_of_context, insufficient]
        detectedProblems:
          type: array
          items: { type: string }
          nullable: true
        notes: { type: string }
        finalVerdict:
          type: string
          enum: [accept, accept_with_caution, reject, irrelevant, needs_better_source]
        createdAt: { type: string, format: date-time }
...
    CreateEvidenceRequest:
      type: object
      required: [title, sourceType]
      properties:
        claimId: { type: string, nullable: true }
        sourceUrl: { type: string, nullable: true }
        title: { type: string }
        publisher: { type: string, nullable: true }
        publishedAt: { type: string, format: date-time, nullable: true }
        sourceType:
          type: string
          enum: [official, academic, government, company, news, benchmark, blog, forum, user_input, unknown]
        excerpt: { type: string, nullable: true }
        summary: { type: string, nullable: true }
        stalenessRisk:
          type: string
          enum: [low, medium, high]
          default: medium
    UpdateEvidenceRequest:
      type: object
      properties:
        claimId: { type: string, nullable: true }
        title: { type: string }
        excerpt: { type: string }
        summary: { type: string }
        reliability:
          type: string
          enum: [pending, high, medium, low, unusable]
        relevance:
          type: string
          enum: [pending, direct, indirect, weak, irrelevant]
        status:
          type: string
          enum: [pending_review, accepted, accepted_with_caution, rejected, irrelevant, needs_better_source]
        stalenessRisk:
          type: string
          enum: [low, medium, high]
    SearchEvidenceRequest:
      type: object
      properties:
        query: { type: string, nullable: true }
        provider:
          type: string
          enum: [manual, web, searxng]
          default: web
        maxResults:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
    AssessEvidenceRequest:
      type: object
      required: [modelIds]
      properties:
        modelIds:
          type: array
          minItems: 1
          items: { type: string }
...
    EvidenceResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Evidence'
    EvidenceListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Evidence'
    EvidenceAssessmentListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceAssessment'
...